Support BSD sed and older bash in test rule.
BSD sed (as shipped with MacOS) does not support `+` nor `\+` in regular expressions.
However it does support `\{1,\}` which does the same thing and is also supported by GNU sed.

Moreover, bash 3.2 (as shipped with MacOS) does not seem to support `source <(...)`, it seems to just do nothing.
Using plain eval seems to do the trick though.