Do not retry updates on conflict (#217)
* Do not retry on conflict
When the reconciliation loops tries to update a CR, and the result is a Conflict, the reconciliation loop should not retry because this error will never succeed. The reconciliation loop needs to be restarted.

Additionally, there was a situation where the resource would be deleted, but the Updater would still try to update the deleted resource, resulting in errors.