workflow/deploy: don't process all branches
In the deploy workflow, using a push branch config of ** results in it
trying to gpg init for branches created by dependabot or anyone/anything
else. This is not what we want - we want those covered by pull request
events.

Change the config to only handle pushes to main (and continue to handle
all tags and all PRs).

Signed-off-by: Andy Goldstein <andy.goldstein@redhat.com>