[Add] Allow specifying helm secret namespace for cluster scope objects
Currently, the helm secret namespace is taken from the namespace of
the object which is being passed to `ActionClientFor`. When the action
client is created for cluster scoped objects, the namespace has to
explicitly set in their metadata. This PR introduces an option to
specify the namespace through the context instead.