pkg/client: add history method to ActionClient
In order to give users of this library more information about the state
of a release, we are adding the `History` method to the `ActionClient`
interface.

Without this, the only information is available via the `ActionClient`
about a release is the latest revision (via the `Get` method). However
this proves insufficient if callers need to know details of release
history, for example if an upgrade failed and callers want to know about
the most recently deployed release.

Signed-off-by: Joe Lanford <joe.lanford@gmail.com>