pkg/client: add Config method to ActionClient (#424)
This exposes action.Configuration of a specific
actionClient instance.

General motivation behind it is for the users to have
an 'escape hatch' access to the underlying dependency
objects that action.Configuration holds, which include
for example:
- release records store
- registry client
- Kubernetes API client

More specific motivation comes from a scenario where
an access to the underlying release store instance
is needed in order to gracefully handle pending helm
releases which were previously interrupted.
For additional context on this, see:
https://github.com/operator-framework/operator-controller/pull/1776

Signed-off-by: Artur Zych <5843875+azych@users.noreply.github.com>
Co-authored-by: Artur Zych <5843875+azych@users.noreply.github.com>